kafka: move schema registry related logic to a separate file in order to distinguish from the upcoming update pertaining kafka rest
Delete unused karapace client creation function since all tests rely on fixtures
Run make format, add yapf to requirement list
Add reformat to test target prerequisites, fail on reformat diff found
Update setup.py entry_points with new main function location
Set yapf package version in line with the same version used in other packages
Run make yapf with newer version
