+++ requirements-dev.txt
@@ -5,3 +5,4 @@ flake8==3.7.9
 pylint==2.4.4
 pytest==5.4.1
 requests==2.23.0
+yapf==0.27.0
